9/9 L5
changed the handle etc to ySu or ySuU removing the previous error and changed the two ySuHandle lines (style left&top) which were. either/both of these could have been causing the xDe and ySu lines to be moving together

i fixed above things, and changed the Su variables to xSuU and ySuU to be able to implement another handle (xSuHandle) and to use different variables (ySuL). i also added the ability to move the ySuHandle in x and y directions. also adding recalculateIntersections() at additional parts of the code to make it recalculate them as things are updated