12/9 L1+2
fixed the error with the xSuHandle not appearing (adding the div to index so it had something to put the handle in). the xSuHandle also moves that part of the line (possibly should rename xSu and ySu handle names)

checking again today the errors with resizing seem to be fixed with the chart sizes that were implemented yesterday.

with the recalculateIntersections() function, we went over it again and fixed it (went back over the math redid it to make sure) and the math was all correct but the computer was treating '+ yDe' as a string so y would be negative (which got fixed when yDe would be added) with a string added to it (e.g. y = -0.5444.8 with the yDe=0.8). this was fixed by adding parseFloat as an else statement in the mousemove function to yDe (and then to all variables to avoid future errors).

i added in ySuL into the first y value of the s->d line in the updateChartData(chart) function which fixed the issues with s->d not being correct, which also fixed the issues with the equilibrium lines not being the middle value.