16/9 L5
fixing the recalulateIntersections() function by redoing m and c for both liens (turning them into variables) and adding them into the xIn and yIn variables.

then adding else if statements to the lines so the handles cannot go past the equilibrium. this may have to get improved as there are small errors with this (very minor)