L1+2 19/9
fixed the form and added to it. i added a min so you had to have 0 or more (like what the graph forces). and added if statements that it has to be equal to the equilibrium. also started (and need to finish) adding defaults for the inputs.

i also added a function and button to reset the graph to the defaults and plan to make this include the inputs.