19/9 8:34pm
i finished up the form and moved it to the summary so the page was laid out better. i also added the defaults for the inputs, and added the defaults to the reset function so there shouldn't be an issue with blank boxes.

im considering how i want to lay out the 'set values' and the equilibrium summary and im not sure whether i should change it