23/9 L5
i fixed the surpluses by making the line go to (xIn, yIn) and then would fill to the y-equilibrium for both of the surpluses.

i also added the additional supply and demand lines making it so they are moved but mirroring the originals. i still need to add in the handles and everything else with them moving.

i also tidied up the code by getting rid of some of the comments (of console.log or the past versions of code)