26/9 L1+2
adding to the additional demand and supply lines, making them move with the handles correctly.

we attempted to get the handles to be hidden when the lines are hidden, but had some issues with it (it could work when you move the handles but not onClick of the legend). this is getting put to the side for now and will be relooked at.